separate client and store code
cmake changes

add missing file

fixes

cmake-format

revert dlmalloc.c rename

clang-format

fix

shuffle more files

update
